+++ src/hdmf/common/resources.py
@@ -88,7 +88,7 @@ class ObjectTable(Table):
     __defaultname__ = 'objects'
 
     __columns__ = (
-        {'name': 'files_idx', 'type': int,
+        {'name': 'files_idx', 'type': (int, np.integer),
          'doc': 'The row idx for the file_object_id in FileTable containing the object.'},
         {'name': 'object_id', 'type': str,
          'doc': 'The object ID for the Container/Data.'},
@@ -119,9 +119,9 @@ class ObjectKeyTable(Table):
     __defaultname__ = 'object_keys'
 
     __columns__ = (
-        {'name': 'objects_idx', 'type': (int, Object),
+        {'name': 'objects_idx', 'type': (int, np.integer, Object),
          'doc': 'The index into the objects table for the Object that uses the Key.'},
-        {'name': 'keys_idx', 'type': (int, Key),
+        {'name': 'keys_idx', 'type': (int, np.integer, Key),
          'doc': 'The index into the keys table that is used to make an external resource reference.'}
     )
 
@@ -134,9 +134,9 @@ class EntityKeyTable(Table):
     __defaultname__ = 'entity_keys'
 
     __columns__ = (
-        {'name': 'entities_idx', 'type': (int, Entity),
+        {'name': 'entities_idx', 'type': (int, np.integer, Entity),
          'doc': 'The index into the EntityTable for the Entity that associated with the Key.'},
-        {'name': 'keys_idx', 'type': (int, Key),
+        {'name': 'keys_idx', 'type': (int, np.integer, Key),
          'doc': 'The index into the KeyTable that is used to make an external resource reference.'}
     )
 
@@ -289,7 +289,7 @@ class HERD(Container):
 
     @docval({'name': 'container', 'type': (str, AbstractContainer),
              'doc': 'The Container/Data object to add or the object id of the Container/Data object to add.'},
-            {'name': 'files_idx', 'type': int,
+            {'name': 'files_idx', 'type': (int, np.integer),
              'doc': 'The file_object_id row idx.'},
             {'name': 'object_type', 'type': str, 'default': None,
              'doc': ('The type of the object. This is also the parent in relative_path. If omitted, '
@@ -317,8 +317,8 @@ class HERD(Container):
         obj = Object(files_idx, container, object_type, relative_path, field, table=self.objects)
         return obj
 
-    @docval({'name': 'obj', 'type': (int, Object), 'doc': 'The Object that uses the Key.'},
-            {'name': 'key', 'type': (int, Key), 'doc': 'The Key that the Object uses.'})
+    @docval({'name': 'obj', 'type': (int, np.integer, Object), 'doc': 'The Object that uses the Key.'},
+            {'name': 'key', 'type': (int, np.integer, Key), 'doc': 'The Key that the Object uses.'})
     def _add_object_key(self, **kwargs):
         """
         Specify that an object (i.e. container and relative_path) uses a key to reference
@@ -327,8 +327,8 @@ class HERD(Container):
         obj, key = popargs('obj', 'key', kwargs)
         return ObjectKey(obj, key, table=self.object_keys)
 
-    @docval({'name': 'entity', 'type': (int, Entity), 'doc': 'The Entity associated with the Key.'},
-            {'name': 'key', 'type': (int, Key), 'doc': 'The Key that the connected to the Entity.'})
+    @docval({'name': 'entity', 'type': (int, np.integer, Entity), 'doc': 'The Entity associated with the Key.'},
+            {'name': 'key', 'type': (int, np.integer, Key), 'doc': 'The Key that the connected to the Entity.'})
     def _add_entity_key(self, **kwargs):
         """
         Add entity-key relationship to the EntityKeyTable.
@@ -599,7 +599,9 @@ class HERD(Container):
             {'name': 'field', 'type': str, 'default': '',
              'doc': ('The field of the compound data type using an external resource.')},
             {'name': 'key', 'type': (str, Key), 'default': None,
-             'doc': 'The name of the key or the Key object from the KeyTable for the key to add a resource for.'},
+             'doc': ('The name of the key or the Key object from the KeyTable for the key to add a resource for. '
+                     'If not provided and ``attribute`` names a scalar string attribute, the value of that '
+                     'attribute is used as the key.')},
             {'name': 'entity_id', 'type': str, 'doc': 'The identifier for the entity at the resource.'},
             {'name': 'entity_uri', 'type': str, 'doc': 'The URI for the identifier at the resource.', 'default': None},
             {'name': 'file',  'type': HERDManager, 'doc': 'The file associated with the container.',
@@ -625,6 +627,21 @@ class HERD(Container):
         entity_id = kwargs['entity_id']
         entity_uri = kwargs['entity_uri']
         file = kwargs['file']
+
+        ##########################################
+        # Default the key from a scalar attribute
+        ##########################################
+        if key is None and attribute is not None:
+            if not isinstance(container, AbstractContainer):
+                msg = ("Cannot default 'key' from attribute '%s' because 'container' is not a "
+                       "Container/Data object. Provide 'key' explicitly." % attribute)
+                raise ValueError(msg)
+            attribute_value = getattr(container, attribute)
+            if not isinstance(attribute_value, str):
+                msg = ("Cannot default 'key' from attribute '%s' because its value is not a single "
+                       "string. Provide 'key' explicitly." % attribute)
+                raise ValueError(msg)
+            key = attribute_value
 
         ##################
         # Set File if None
@@ -673,7 +690,6 @@ class HERD(Container):
             # for this entity and key combination.
             check_entity_key = True
             if entity_uri is not None:
-                entity_uri = entity.entity_uri
                 msg = 'This entity already exists. Ignoring new entity uri'
                 warn(msg, stacklevel=3)
 
@@ -893,7 +909,9 @@ class HERD(Container):
             entity_key_row_idx = self.entity_keys.which(keys_idx=key_idx)
             for row_idx in entity_key_row_idx:
                 entity_idx = self.entity_keys['entities_idx', row_idx]
-                entities.append(self.entities.__getitem__(entity_idx))
+                # coerce the row to a tuple so a read-back numpy structured-array row
+                # (numpy.void) expands into columns the same as an in-memory list row
+                entities.append(tuple(self.entities[entity_idx]))
         df = pd.DataFrame(entities, columns=['entity_id', 'entity_uri'])
         return df
 
@@ -947,8 +965,9 @@ class HERD(Container):
         result_df.reset_index(inplace=True, drop=True)
         # ADD files
         file_id_col = []
+        files_df = self.files.to_dataframe()
         for idx in result_df['files_idx']:
-            file_id_val = self.files.to_dataframe().iloc[int(idx)]['file_object_id']
+            file_id_val = files_df.iloc[int(idx)]['file_object_id']
             file_id_col.append(file_id_val)
 
         result_df['file_object_id'] = file_id_col
@@ -969,6 +988,50 @@ class HERD(Container):
             result_df.columns = pd.MultiIndex.from_tuples(column_labels)
         # return the result
         return result_df
+
+    def __flattened_dataframe_or_none(self):
+        """Return the flattened ``to_dataframe()`` view, or None when there are no references.
+
+        ``to_dataframe`` raises when the HERD holds no object-key relationships and may fail if the
+        backing file is closed. The repr methods use this helper so they never raise on display.
+        """
+        if len(self.object_keys) == 0:
+            return None
+        try:
+            return self.to_dataframe()
+        except Exception:
+            return None
+
+    def __summary_line(self):
+        """Return a one-line summary of the table sizes."""
+        return ("%d key(s), %d entity(ies), %d object(s), %d file(s)"
+                % (len(self.keys), len(self.entities), len(self.objects), len(self.files)))
+
+    def __repr__(self):
+        cls = self.__class__
+        template = "%s %s.%s at 0x%d" % (self.name, cls.__module__, cls.__name__, id(self))
+        template += "\n  " + self.__summary_line()
+        df = self.__flattened_dataframe_or_none()
+        if df is not None and len(df) > 0:
+            template += "\n" + repr(df)
+        return template
+
+    def _repr_html_(self):
+        """Generate an HTML representation that surfaces the references as a flattened table."""
+        header_text = self.name if self.name == self.__class__.__name__ else \
+            f"{self.name} ({self.__class__.__name__})"
+        html_repr = self.css_style + self.js_script
+        html_repr += "<div class='container-wrap'>"
+        html_repr += f"<div class='container-header'><div class='xr-obj-type'><h3>{header_text}</h3></div></div>"
+        html_repr += self._closed_file_warning_html()
+        html_repr += f"<p class='container-fields'>{self.__summary_line()}</p>"
+        df = self.__flattened_dataframe_or_none()
+        if df is None or len(df) == 0:
+            html_repr += "<p class='container-fields'>No external resource references.</p>"
+        else:
+            html_repr += df.to_html()
+        html_repr += "</div>"
+        return html_repr
 
     @docval({'name': 'path', 'type': str, 'doc': 'The path to the zip file.'})
     def to_zip(self, **kwargs):
@@ -1001,12 +1064,14 @@ class HERD(Container):
         return directory
 
     @classmethod
-    @docval({'name': 'path', 'type': str, 'doc': 'The path to the zip file.'})
+    @docval({'name': 'path', 'type': str, 'doc': 'The path to the zip file.'},
+            {'name': 'type_map', 'type': TypeMap, 'default': None,
+             'doc': 'The TypeMap to use for the returned HERD. If None, the default TypeMap is used.'})
     def from_zip(cls, **kwargs):  # noqa: C901
         """
         Method to read in zipped tsv files to populate HERD.
         """
-        zip_file = kwargs['path']
+        zip_file, type_map = popargs('path', 'type_map', kwargs)
         directory = cls.get_zip_directory(zip_file)
 
         with zipfile.ZipFile(zip_file, 'r') as zip:
@@ -1077,12 +1142,13 @@ class HERD(Container):
                 msg = "Key Index out of range in EntityKeyTable. Please check for alterations."
                 raise ValueError(msg)
 
-        er = HERD(
+        er = cls(
             files=files,
             keys=keys,
             entities=entities,
             entity_keys=entity_keys,
             objects=objects,
-            object_keys=object_keys
+            object_keys=object_keys,
+            type_map=type_map,
         )
         return er